requirements.txt
# Web framework
flask >= 3.1.3

# Database requirements
Flask-SQLAlchemy==3.1.1
alembic==1.18.3
psycopg2-binary>=2.9.5

# Environment variable management
python-dotenv==1.2.1

# Testing requirements
pytest==9.0.2

# LLMs requirements
# Semantic Kernel for orchestration with function calling support
semantic-kernel>=1.18.0
# OpenAI API (for main orchestration and Generator Agent)
openai>=1.0.0
# Anthropic Claude API (for Detector Agent)
anthropic>=0.18.0
# Google Gemini API (for Judge Agent)
google.genai==1.39.0
# Environment variables
python-dotenv>=1.2.1
# Async support (included with other packages, but explicit for clarity)
aiohttp>=3.9.0
# Calculate LLM cost
tokencost>=0.1.26

#OpenAI Agents framework for more advanced agent management
openai-agents>=0.10.0
openai-agents[litellm]